in debian/control: 	added paella-machine-installer package 	paella-nfs-installer now depends on paella-machine-installer 	machine installer depends on grub added paella-machine-installer script added paella-setup-machine-installer script
added methods to save install logfile to installed system in installer scripts
	
started disk manager in paella-management:
	made DiskDoc class in kde.docgen.machine
	made DiskView class in kde.machine.viewbrowser
	made DiskManager class in kde.machine.main
added self.current attribute to BaseProcessor to keep track of current process
in installer.chroot:
	added _install_finished flag
	added mtypedata
	keep backup of installer apt-sources when installing final apt-sources
	ready_base_for_install now updates apt-sources
	added save_logfile_in_target method to save install logs
in installer.trait:
	don't call apt with empty package list in install/remove packages
	add some code to handle the 128 errors that spuriously pop up
	  when running the paella-debconf.sh script.  Now we re-run the script
	  when a 128 is returned until it returns 0.  The method is set to fail after
	  20 failed attempts
in installer.machine:
	added some function decorators for the machine installer
	  while i like the idea i have for using decorators as assertions,
	  the implementation may need some more work, as tracebacks
	  produce a long chain of wrapper functions (which is ugly)
	added flags:
	  _target_mounted
	  _disks_setup
	fix mtypedata when setting machine
	remove old code
in installer.machinehelper:
	add log member to helper
	install_kernel method prepares grub first, the code is pretty hacky
	  and not complete yet, although it should work in most cases when
	  installing to /dev/hda
	remove code that is now in chroot installer
remove echo from installer.util.filesystem

changes while testing:
  * save install log in root/paella on installed machine
  * fix update-grub to assume yes
  * added simple grub handling to install_kernel helper
  * adding log attribute to machine installer helper
  * update kernel install method
  * call ready_base_for_install for chroot installer from machine installer
  * quick syntax error bugfix
  * still working with debconf scripts
  * try to fix debconf scripts
  * don't call apt-get remove with empty package list
  * fix imports in machinehelper
  * quick bugfix
  * make sure logdir exists
  * fix string substitution in disklogpath
  * make disklog path
  * need to get rid of echo from useless.base.util
  * still testing new paella-machine-installer package
  * still testing new paella-machine-installer package
  * test new paella-machine-installer package
  * make sure debconf.sh exists before executing
  * add another hacky way to use debconf
  * fix trait installer logging
  * fixing logging manipulations
  * more logging manipulations
  * fix logdir name
  * test new logging class
  * try to fix logfile for running scripts
  * fix arg in make_script in TraitInstallerHelper
  * fix set_logfile method in chroot installer
  * more log info in chroot installer
  * implement install method in ChrootInstaller
  * fix args to TraitInstaller in ProfileInstaller
  * set logfile in paella-chroot-install script
  * add oldinstaller to paella-installer package
  * starting to actually make the new installers
  * using new runlog function, testing on machine installer
  * some more work
  * the logfile code really needs a lot of work
  * more work with logfile and chroot installer
  * fix config lookup in paella-setup-installer
  * fix logfile in chroot installer class
  * default logfile with paella-chroot-install
  * more testing

+++ src/paella/installer/trait.py
@@ -69,24 +69,31 @@ class TraitInstallerHelper(object):
         #self.log.info('trait set to %s' % trait)
 
     # helper to run commands in a chroot on the target
-    def chroot(self, command):
-        return runlog('chroot %s %s' % (self.target, command))
+    def chroot(self, command, failure_suppressed=False):
+        cmd = 'chroot %s %s' % (self.target, command)
+        return runlog(cmd, failure_suppressed=failure_suppressed)
         
     def remove_packages(self, packages):
-        packages_arg = ' '.join(packages)
-        command = 'apt-get -y remove %s' % packages_arg
-        self.chroot(command)
+        if packages:
+            packages_arg = ' '.join(packages)
+            command = 'apt-get -y remove %s' % packages_arg
+            self.chroot(command)
+        else:
+            self.log.info('No packages to remove')
         
     def install_packages(self, packages, unauthenticated=False):
-        packages_arg = ' '.join(packages)
-        opts = ''
-        if unauthenticated:
-            opts = '--allow-unauthenticated'
-        command = 'apt-get -y %s install %s' % (opts, packages_arg)
-        stmt = 'install command for trait %s is:  %s' % (self.trait, command)
-        self.log.info(stmt)
-        self.chroot(command)
-
+        if packages:
+            packages_arg = ' '.join(packages)
+            opts = ''
+            if unauthenticated:
+                opts = '--allow-unauthenticated'
+            command = 'apt-get -y %s install %s' % (opts, packages_arg)
+            stmt = 'install command for trait %s is:  %s' % (self.trait, command)
+            self.log.info(stmt)
+            self.chroot(command)
+        else:
+            self.log.info('No packages to install')
+            
     def remove_cached_debs(self):
         archives = self.target / 'var/cache/apt/archives'
         partial = archives / 'partial'
@@ -287,7 +294,23 @@ class TraitInstaller(BaseInstaller):
             # a hacky way to configure debconf
             if t.template == 'root/paella-debconf.sh':
                 self.log.info('Running paella-debconf.sh script ...')
-                self.helper.chroot('/root/paella-debconf.sh')
+                retval = self.helper.chroot('/root/paella-debconf.sh', failure_suppressed=True)
+                # sometimes this script returns 128 and can't be reproduced reliably
+                # so we try running it again, until we have no 128 returned
+                if retval == 128:
+                    count = 0
+                    while retval == 128:
+                        count += 1
+                        if count < 20:
+                            self.log.info('/root/paella-debconf.sh returned 128, running again')
+                            # this time we will fail on 128
+                            retval = self.helper.chroot('/root/paella-debconf.sh',
+                                                        failure_suppressed=True)
+                        else:
+                            raise InstallDebconfError, 'Too many attempts to run debconf script'
+                elif retval and retval != 128:
+                    msg = "problem running debconf script, returned %d" % retval
+                    raise InstallDebconfError, msg
                 # rename script
                 filename = self.target / 'root/paella-debconf.sh'
                 newname = self.target / 'root/%s-paella-debconf.sh' % self.helper.trait
